Rename to yank_grid_measure_buffer_action; Move actions
