+++ include/fullscore/actions/yank_grid_measure_to_buffer_action.h
@@ -10,15 +10,15 @@ namespace Measure { class Basic; class Base; }
 
 namespace Action
 {
-   class YankMeasureToBuffer : public Base
+   class YankGridMeasureToBuffer : public Base
    {
    private:
       Measure::Basic *yank_measure_buffer;
       Measure::Base *source_measure;
 
    public:
-      YankMeasureToBuffer(Measure::Basic *yank_measure_buffer, Measure::Base *source_measure);
-      ~YankMeasureToBuffer();
+      YankGridMeasureToBuffer(Measure::Basic *yank_measure_buffer, Measure::Base *source_measure);
+      ~YankGridMeasureToBuffer();
 
       bool execute();
    };
